PythonProject/requirements.txt
grpcio
grpcio-tools
python-dotenv
psycopg2-binary
SQLAlchemy
